Refactor search tactics
