+++ dynast/search/tactic/__init__.py
@@ -1,0 +1,4 @@
+from .random import RandomSearch
+from .distributed import LINASDistributed, RandomSearchDistributed
+from .evolutionary import Evolutionary
+from .linas import LINAS
