+++ dynast/search/tactic/distributed/__init__.py
@@ -1,0 +1,2 @@
+from .random import RandomSearchDistributed
+from .linas import LINASDistributed
